minor review comment fixed
